+++ mayan/apps/documents/forms/document_type_forms.py
@@ -3,9 +3,42 @@ from django.utils.translation import ugettext_lazy as _
 
 from mayan.apps.acls.models import AccessControlList
 
+from ..classes import BaseDocumentFilenameGenerator
 from ..models import DocumentType, DocumentTypeFilename
 
-__all__ = ('DocumentTypeFilteredSelectForm', 'DocumentTypeFilenameForm_create')
+__all__ = (
+    'DocumentTypeFileGeneratorForm', 'DocumentTypeFilteredSelectForm',
+    'DocumentTypeFilenameForm_create'
+)
+
+
+class DocumentTypeFileGeneratorForm(forms.ModelForm):
+    class Meta:
+        fields = (
+            'filename_generator_backend',
+            'filename_generator_backend_arguments'
+        )
+        model = DocumentType
+        widgets = {
+            'filename_generator_backend': forms.widgets.Select(
+                choices=BaseDocumentFilenameGenerator.get_choices()
+            )
+        }
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.fields[
+            'filename_generator_backend'
+        ].choices = BaseDocumentFilenameGenerator.get_choices()
+
+
+class DocumentTypeFilenameForm_create(forms.ModelForm):
+    """
+    Model class form to create a new document type filename
+    """
+    class Meta:
+        fields = ('filename',)
+        model = DocumentTypeFilename
 
 
 class DocumentTypeFilteredSelectForm(forms.Form):
@@ -42,12 +75,3 @@ class DocumentTypeFilteredSelectForm(forms.Form):
             widget=widget_class(attrs={'class': 'select2', 'size': 10}),
             **extra_kwargs
         )
-
-
-class DocumentTypeFilenameForm_create(forms.ModelForm):
-    """
-    Model class form to create a new document type filename
-    """
-    class Meta:
-        fields = ('filename',)
-        model = DocumentTypeFilename