+++ mayan/apps/documents/models/document_type_models.py
@@ -7,7 +7,10 @@ from django.utils.translation import ugettext_lazy as _
 
 from mayan.apps.acls.models import AccessControlList
 from mayan.apps.common.literals import TIME_DELTA_UNIT_CHOICES
+from mayan.apps.common.serialization import yaml_load
+from mayan.apps.common.validators import YAMLValidator
 
+from ..classes import BaseDocumentFilenameGenerator
 from ..events import event_document_type_created, event_document_type_edited
 from ..literals import DEFAULT_DELETE_PERIOD, DEFAULT_DELETE_TIME_UNIT
 from ..managers import DocumentTypeManager
@@ -47,6 +50,20 @@ class DocumentType(models.Model):
         blank=True, choices=TIME_DELTA_UNIT_CHOICES,
         default=DEFAULT_DELETE_TIME_UNIT, max_length=8, null=True,
         verbose_name=_('Delete time unit')
+    )
+    filename_generator_backend = models.CharField(
+        default=BaseDocumentFilenameGenerator.get_default(), help_text=_(
+            'The class responsible for producing the actual filename used '
+            'to store the uploaded documents.'
+        ), max_length=224, verbose_name=_('Filename generator backend')
+    )
+    filename_generator_backend_arguments = models.TextField(
+        blank=True, help_text=_(
+            'The arguments for the filename generator backend as a '
+            'YAML dictionary.'
+        ), validators=[YAMLValidator()], verbose_name=_(
+            'Filename generator backend arguments'
+        )
     )
 
     objects = DocumentTypeManager()
@@ -91,6 +108,19 @@ class DocumentType(models.Model):
         )
 
         return queryset.count()
+
+    def get_upload_filename(self, instance, filename):
+        generator_klass = BaseDocumentFilenameGenerator.get(
+            name=self.filename_generator_backend
+        )
+        generator_instance = generator_klass(
+            **yaml_load(
+                stream=self.filename_generator_backend_arguments or '{}'
+            )
+        )
+        return generator_instance.upload_to(
+            instance=instance, filename=filename
+        )
 
     def natural_key(self):
         return (self.label,)