+++ mayan/apps/documents/tests/test_models.py
@@ -265,6 +265,15 @@ class DocumentTypeModelTestCase(GenericDocumentTestCase):
     def test_method_get_absolute_url(self):
         self.assertTrue(self.test_document_type.get_absolute_url())
 
+    def test_original_filename_backend(self):
+        self.test_document_type.filename_generator_backend = 'original'
+        self.test_document_type.save()
+        self._upload_test_document()
+        self.assertEqual(
+            self.test_document_filename,
+            self.test_document.latest_version.file
+        )
+
 
 class DuplicatedDocumentsTestCase(GenericDocumentTestCase):
     def test_duplicates_after_delete(self):